continue fixing all examples for use on mingw with qtcreator

+++ class-01-basic-game-dev/04-0-render-basic/01_line_render_main.cxx
@@ -6,8 +6,8 @@ int main(int, char**)
     const color white = { 255, 255, 255 };
     const color green = { 0, 255, 0 };
 
-    size_t width  = 320;
-    size_t height = 240;
+    int32_t width  = 320;
+    int32_t height = 240;
 
     canvas image(width, height);
